SPORTS-48 Rename cardv2 to treeCard
