+++ frontend/src/components/tournament/tree/treeCard/Player.js
