+++ frontend/src/components/tournament/tree/treeCard/PlayerPhoto.js
